+++ Sources/VComponents/API/Deprecations.swift
@@ -6,3 +6,35 @@
 //
 
 import SwiftUI
+
+@available(tvOS, unavailable)
+@available(visionOS, unavailable)
+extension VStretchedButtonAppearance {
+    @available(*, unavailable, message: "Use 'labelSpacingType' instead")
+    public var labelSpacing: CGFloat {
+        get { fatalError() }
+        set { fatalError() }
+    }
+}
+
+@available(tvOS, unavailable)
+@available(watchOS, unavailable)
+@available(visionOS, unavailable)
+extension VLoadingStretchedButtonAppearance {
+    @available(*, unavailable, message: "Use 'labelSpacingType' instead")
+    public var labelSpacing: CGFloat {
+        get { fatalError() }
+        set { fatalError() }
+    }
+}
+
+@available(tvOS, unavailable)
+@available(watchOS, unavailable)
+@available(visionOS, unavailable)
+extension VStretchedToggleButtonAppearance {
+    @available(*, unavailable, message: "Use 'labelSpacingType' instead")
+    public var labelSpacing: CGFloat {
+        get { fatalError() }
+        set { fatalError() }
+    }
+}