+++ Sources/VComponents/Presentation/Components/Buttons/Stretched Button (Loading)/VLoadingStretchedButtonAppearance.swift
@@ -89,10 +89,10 @@ public struct VLoadingStretchedButtonAppearance {
     /// Label text and label image placement.
     public var labelTextAndLabelImagePlacement: TextAndImagePlacement = .imageAndText
 
-    /// Label spacing.
+    /// Label spacing type.
     ///
     /// Applicable only if `init` with multiple components is used.
-    public var labelSpacing: CGFloat = 8
+    public var labelSpacingType: TextAndImageSpacingType = .fixed(spacing: 8)
 
     /// Label pressed scale.
     public var labelPressedScale: CGFloat = {
@@ -223,6 +223,14 @@ public struct VLoadingStretchedButtonAppearance {
             switch self {
             case .fixed: false
             case .stretched: true
+            }
+        }
+        
+        var centeredElementMaxWidth: CGFloat? {
+            if hasFlexibleSpacing {
+                CGFloat.infinity
+            } else {
+                nil
             }
         }