+++ Sources/VComponents/Presentation/Components/Buttons/Stretched Button/VStretchedButtonAppearance.swift
@@ -88,10 +88,10 @@ public struct VStretchedButtonAppearance {
     /// Label text and label image placement.
     public var labelTextAndLabelImagePlacement: TextAndImagePlacement = .imageAndText
 
-    /// Label spacing.
+    /// Label spacing type.
     ///
     /// Applicable only if `init` with multiple components is used.
-    public var labelSpacing: CGFloat = 8
+    public var labelSpacingType: TextAndImageSpacingType = .fixed(spacing: 8)
 
     /// Label pressed scale.
     public var labelPressedScale: CGFloat = {